refactor: modularize backend for improved maintainability
- Split monolithic 1337-line main.py into focused modules
- Reduce main.py to 69 lines (96% reduction)
- Create service layer (LLM, RLM, image, logging services)
- Add API versioning structure (v1 endpoints)
- Separate middleware (security headers, CORS)
- Add utilities (state management, error handlers, security)
- Implement Pydantic schemas for request validation
- Fix config endpoint field names (models → available_models)
- Support both @image and /image commands for image generation
- Update README with new architecture documentation
- Maintain 100% backward compatibility with all features
