+++ app/__init__.py
@@ -1,0 +1,3 @@
+"""Package initializer for app."""
+
+__version__ = "0.3.0"
