+++ app/api/schemas/__init__.py
@@ -1,0 +1,5 @@
+"""Pydantic models for request/response validation."""
+
+from .chat import ChatRequest, RLMPasscodeRequest
+
+__all__ = ["ChatRequest", "RLMPasscodeRequest"]
