+++ app/dependencies.py
@@ -1,0 +1,7 @@
+"""FastAPI dependencies and dependency injection."""
+
+# Currently no dependencies needed, but this file is reserved for future use
+# Examples of what might go here:
+# - Database session dependencies
+# - Authentication dependencies
+# - Rate limiting dependencies
