+++ app/middleware/__init__.py
@@ -1,0 +1,1 @@
+"""Middleware package for TinyChat."""
