+++ app/middleware/security.py
@@ -1,0 +1,57 @@
+"""Security middleware for HTTP headers and CORS."""
+
+from fastapi import Request
+from fastapi.middleware.cors import CORSMiddleware
+from fastapi.middleware.trustedhost import TrustedHostMiddleware
+
+from app.config import Settings
+
+
+def setup_security_middleware(app):
+    """Add security middleware to the FastAPI app."""
+    
+    # Add trusted host middleware if hosts are configured
+    if Settings.ALLOWED_HOSTS != ["*"]:
+        app.add_middleware(TrustedHostMiddleware, allowed_hosts=Settings.ALLOWED_HOSTS)
+    
+    # Add CORS middleware
+    app.add_middleware(
+        CORSMiddleware,
+        allow_origins=Settings.ALLOWED_ORIGINS,
+        allow_credentials=True,
+        allow_methods=["GET", "POST", "DELETE"],
+        allow_headers=["*"],
+    )
+
+
+async def add_security_headers(request: Request, call_next):
+    """
+    Middleware to add security headers to all HTTP responses.
+    
+    Adds headers for:
+    - X-Content-Type-Options: Prevent MIME sniffing
+    - X-Frame-Options: Prevent clickjacking
+    - X-XSS-Protection: Enable browser XSS filters
+    - Strict-Transport-Security: Enforce HTTPS
+    - Content-Security-Policy: Restrict resource loading
+    
+    Args:
+        request: The incoming request
+        call_next: The next middleware/handler in the chain
+        
+    Returns:
+        Response with added security headers
+    """
+    response = await call_next(request)
+    response.headers["X-Content-Type-Options"] = "nosniff"
+    response.headers["X-Frame-Options"] = "DENY"
+    response.headers["X-XSS-Protection"] = "1; mode=block"
+    response.headers["Strict-Transport-Security"] = "max-age=31536000; includeSubDomains"
+    response.headers["Content-Security-Policy"] = (
+        "default-src 'self'; "
+        "script-src 'self' 'unsafe-inline' https://cdn.jsdelivr.net https://cdnjs.cloudflare.com; "
+        "style-src 'self' 'unsafe-inline' https://cdn.jsdelivr.net https://cdnjs.cloudflare.com; "
+        "font-src 'self' https://cdn.jsdelivr.net; "
+        "img-src 'self' data:"
+    )
+    return response
